Refactored PanZoom in new _panzoom.py module.
